+++ app/src/main/Kotlin/LibraryItems.kt
@@ -1,0 +1,10 @@
+package classes
+
+abstract class LibraryItems(
+    open val id: Int,
+    open val name: String,
+    open var isAvailable: Boolean
+) {
+    abstract fun getShortInfo() : String
+    abstract fun getAllInfo() : String
+}
