+++ app/src/main/Kotlin/ReadInLibrary.kt
@@ -1,0 +1,5 @@
+package interfaces
+
+interface ReadInLibrary {
+    fun readInLibrary()
+}
