+++ app/src/main/Kotlin/ReturnItem.kt
@@ -1,0 +1,5 @@
+package interfaces
+
+interface ReturnItem {
+    fun returnItem()
+}
