+++ app/src/main/Kotlin/TakeHome.kt
@@ -1,0 +1,5 @@
+package interfaces
+
+interface TakeHome {
+    fun takeHome()
+}
